+++ src/index.ts
@@ -2,4 +2,8 @@ export * from "./client";
 export * from "./codec";
 export * from "./errors";
 export type * from "./types";
-export { SECURE_TRANSFER_CONTRACT } from "./types";
+export {
+  SECURE_TRANSFER_CONTRACT,
+  SECURE_TRANSFER_REVOCATION_CONTRACT,
+  SECURE_TRANSFER_UPLOAD_RECEIPT_CONTRACT,
+} from "./types";